utils/weibo/decode.py
import json
import time
import requests
import base64
import execjs
import io
import cv2
import os
import numpy as np
from PIL import Image
import math
import random
import re
from lxml import etree
import execjs.runtime_names
default = execjs.get(execjs.runtime_names.PhantomJS)
print(default.name)
x=default.eval("1+2")
print(x)