requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
backcall==0.1.0
baidu-aip==2.2.5.1
beautifulsoup4==4.6.0
billiard==3.5.0.4
bs4==0.0.1
celery==4.2.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
cryptography==2.2.2
decorator==4.3.0
Django==2.0.7
django-cors-headers==2.3.0
django-filter==1.1.0
django-redis==4.9.0
django-redis-cache==1.7.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
EasyProcess==0.2.3
idna==2.7
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.12.1
kombu==4.2.1
numpy==1.15.2
opencv-python==3.4.3.18
parso==0.3.1
pickleshare==0.7.5
Pillow==5.2.0
prompt-toolkit==2.0.4
pycparser==2.18
PyExecJS==1.5.1
Pygments==2.2.0
PyJWT==1.6.4
PyMySQL==0.9.2
PySocks==1.6.8
pytz==2018.5
PyVirtualDisplay==0.2.1
redis==2.10.6
requests==2.19.1
requests-toolbelt==0.8.0
selenium==3.14.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
vine==1.1.4
wcwidth==0.1.7
